policy/pyproject.toml
[project]
name = "llm_policy"
version = "0.1.0"
description = "Policy engine for gating LLM capabilities (e.g., extract enablement) based on eval artifacts and telemetry."
requires-python = "==3.12.*"

dependencies = [
  # YAML parsing for thresholds + models config
  "pyyaml>=6.0.1",

  # Typed configs / decision objects
  "pydantic>=2.7",

  # If/when you read telemetry from llm-server admin endpoints
  "httpx>=0.27",

  # Convenience parsing / filesystem
  "python-dotenv>=1.0.0",
]

[project.scripts]
llm-policy = "llm_policy.cli:main"

[project.optional-dependencies]
test = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5",
  "coverage>=7",
]

lint = [
  "ruff>=0.5",
  "black>=24",
]

dev = [
  "ipython>=8",
  "rich>=13.7",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["llm_policy*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
llm_policy = ["**/*.yaml", "**/*.yml", "**/*.json"]

[tool.pytest.ini_options]
addopts = "-q --cov=llm_policy --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.black]
line-length = 100
target-version = ["py312"]